// src/hcamain/src/temp/component-library/components/CQCBlock/CQCBlock.types.ts
import { ReactNode } from 'react';
export interface CQCBlockProps {
  logo: {
    dark: ReactNode | JSX.Element;
    light: ReactNode | JSX.Element;
  };
  title?: string;
  text?: string;
  icon: ReactNode | JSX.Element;
  length?: 'short' | 'long';
  rating?: 'Outstanding' | 'Good' | 'Requires improvement' | 'Inadequate';
  link: ReactNode | JSX.Element;
}